Some fixes for screenevalsummary
Fixed a few things displaying improperly when switching between FA+ and normal FA being displayed.

The judgement numbers are still not displaying properly. So if a player plays a song on page 1 with FA+ enabled their fantastic count will be shown again on page 2 in the same slot as the song on page 1 had FA+ enabled on top of the other judgements already there for that song.