Remove redundant code.
This appears later in the file and doesn't need to be here twice. (this is actually the commit name/description of the previous commit I just messed up on the push order and I'm too lazy/dumb to figure out how to undo it so I'm just gonna let yall figure it out LOL).