Always have mouse input enabled + mouse preview music.
We always want mouse input enabled now as we are veering into more and more mouse specific features. One of which is in this commit. You can now click anywhere on the density graph to hear the preview music at that part of the song! You could also click at the beginning of the song to hear the entire thing if you wanted.