More Tag Menu Fixes.
If you happen to be using the beta version of tags already you'll want to edit your tags file(s) so that all the lines for songs and groups have a "/" at the start of them.

Added "/" to the start of every line that isn't a tag so that if a pack starts with a "#" it wont get mistaken for a tag.

Also fixed some silly logic issues with the Tag Menu AHHHHHH