Add basic "Ghost Window" functionality.
Similar to FA+ in that you get judgment feedback, but it has zero impact on normal score or Ex score.

Currently you can choose any value from 1-22ms (or 1-14ms if you have FA+ enabled).
I plan to expand upon this to maybe reflect on step statistics judgments (along with a label of the ms) and maybe an extra evaluation pane in the future.